+++ src/main/java/ru/sova/educationapp/EducationalSystemApp/repositories/StudentRepository.java
@@ -9,7 +9,7 @@ import ru.sova.educationapp.EducationalSystemApp.models.VerificationWork;
 import java.util.List;
 
 @Repository
-public interface StudentRepository extends JpaRepository<Student, Integer> {
+public interface StudentRepository extends JpaRepository<Student, Long> {
 
     List<Student> findAllByVerificationWorksContains(VerificationWork verificationWork);
