+++ src/test/java/ru/sova/educationapp/EducationalSystemApp/services/TaskServiceTest.java
@@ -24,7 +24,7 @@ import static org.mockito.Mockito.times;
 
 public class TaskServiceTest {
 
-    private static final int ID = 1;
+    private static final long ID = 1;
 
     @Mock
     private TaskRepository taskRepository;
